+++ src/main/java/com/sonnh/retailmanagerv2/dto/request/admin/PromotionUpdateReqDto.java
@@ -4,6 +4,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
@@ -12,7 +13,7 @@ import java.util.UUID;
 @NoArgsConstructor
 public class PromotionUpdateReqDto implements Serializable {
     private String name;
-    private Double discountPercent;
+    private BigDecimal discountPercent;
     private LocalDateTime startDate;
     private LocalDateTime endDate;
     private List<ProductDto> productDtoList;
