+++ src/main/java/com/sonnh/retailmanagerv2/dto/response/admin/ProductAndStoreListResDto.java
@@ -9,6 +9,7 @@ import lombok.NoArgsConstructor;
 import lombok.Value;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
@@ -21,7 +22,7 @@ import java.util.UUID;
 public class ProductAndStoreListResDto implements Serializable,PromotionDetailResDto {
     private UUID promotionId;
     private String name;
-    private Double discountPercent;
+    private BigDecimal discountPercent;
     private LocalDateTime startDate;
     private LocalDateTime endDate;
     private PromotionStatus status;
